feat: 쿠폰 유효성 검사 로직을 getValidCoupon 유틸리티 함수로 분리
