+++ src/domains/orderConfirm/hooks/useCoupons.ts
@@ -1,65 +1,10 @@
 import { useEffect, useState } from "react";
-import { CartItemTypes } from "../../shopping-cart/types/cartItem";
 import { getCouponItems } from "../api/coupon";
-import { CouponCode, CouponCodes, CouponType } from "../types/coupon";
-import {
-  validateDate,
-  validateMinimumAmount,
-  validateTime,
-} from "./utils/validateCoupons";
+import { CouponType } from "../types/coupon";
 
 export function useCoupons() {
   const [coupons, setCoupons] = useState<CouponType[]>([]);
 
-  const today = new Date();
-
-  /**
-   * 쿠폰이 유효한지 확인하는 함수
-   * @param orderPrice 주문 금액
-   * @param twoPlusOneApplicableItems 2+1 적용 가능한 아이템 목록
-   * @returns <Record<CouponCode, boolean>> 쿠폰 코드와 유효성 여부를 매핑한 객체
-   */
-  const isValidCoupon = ({
-    orderPrice,
-    twoPlusOneApplicableItems,
-  }: {
-    orderPrice: number;
-    twoPlusOneApplicableItems: CartItemTypes[];
-  }): Record<CouponCode, boolean> => {
-    return Object.values(CouponCodes).reduce((acc, code) => {
-      return {
-        ...acc,
-        [code]: validateCoupon(code, orderPrice, twoPlusOneApplicableItems),
-      };
-    }, {} as Record<CouponCode, boolean>);
-  };
-
-  const validateCoupon = (
-    couponCode: CouponCode,
-    orderPrice: number,
-    selectedCartItems: CartItemTypes[]
-  ) => {
-    const couponItem = coupons.find((item) => item.code === couponCode);
-    if (!couponItem) return false;
-
-    if (couponItem.code === "BOGO" && selectedCartItems.length === 0)
-      return false;
-
-    return (
-      validateDate({ expirationDate: couponItem.expirationDate, today }) &&
-      (!couponItem.availableTime ||
-        validateTime({ availableTime: couponItem.availableTime, today })) &&
-      (!couponItem.minimumAmount ||
-        validateMinimumAmount({
-          minimumAmount: couponItem.minimumAmount,
-          orderPrice,
-        }))
-    );
-  };
-
-  /**
-   * 쿠폰 목록을 가져오는 함수
-   */
   useEffect(() => {
     (async () => {
       const data = await getCouponItems();
@@ -68,7 +13,6 @@ export function useCoupons() {
   }, []);
 
   return {
-    isValidCoupon,
     coupons,
   };
 }
